apps: Merge auracast and auracast_usb applications
This adds new AUDIO_USB syscfg option to disable the USB audio support
in auracast_usb application.
This fixes code duplication, so that auracast_usb has been renamed
to auracast, where previous auracast application has been removed.
