+++ apps/auracast/pkg.yml
@@ -32,8 +32,17 @@ pkg.deps:
     - nimble/host/store/config
     - nimble/host/audio/services/auracast
     - "@apache-mynewt-core/kernel/os"
+    - "@apache-mynewt-core/sys/config"
     - "@apache-mynewt-core/sys/console"
     - "@apache-mynewt-core/sys/log"
     - "@apache-mynewt-core/sys/stats"
     - "@apache-mynewt-core/sys/sysinit"
     - "@apache-mynewt-core/sys/id"
+
+pkg.deps.AUDIO_USB:
+    - ext/liblc3
+    - ext/libsamplerate
+    - "@apache-mynewt-core/hw/usb/tinyusb"
+
+pkg.init:
+    audio_init: 402
