+++ apps/auracast/syscfg.yml
@@ -16,13 +16,64 @@
 # under the License.
 
 syscfg.defs:
-    BROADCASTER_CHAN_NUM: 2
-    BROADCASTER_BROADCAST_NAME: '"NimBLE Auracast"'
-    BROADCASTER_STOP_BUTTON:
+    AUDIO_USB:
+        description: Enable USB audio output device driver.
+        value: 1
+        restrictions:
+            # For now, LC3 encoded audio data, with 48kHz sample rate is supported only.
+            - '(LC3_SAMPLING_FREQ == 48000) if 0'
+            - '(LC3_FRAME_DURATION == 10000) is 0'
+
+    ISO_HCI_FEEDBACK:
+        description: Enable HCI feedback for resampler. This reduces jitter between USB and ISO.
+        value: 0
+
+    AURACAST_CHAN_NUM: 2
+
+    AURACAST_STOP_BUTTON:
         description: >
-            Button number for Broadcast Stop action.
+            Button number for Auracast Stop action.
             Negative value if disabled.
         value: BUTTON_3
+
+    BROADCAST_ID:
+        description: Broadcast ID value, will use random number if 0
+        value: 0x000000
+    BROADCAST_NAME:
+        description: Broadcast name
+        value: '"NimBLE Auracast"'
+    BROADCAST_CODE:
+        description: Broadcast code used for encrpytion
+        value: '"listen2nimble"'
+
+    LC3_FRAME_DURATION:
+        description: LC3 frame duration
+        value: 10000
+    LC3_SAMPLING_FREQ:
+        description: LC3 sampling frequency
+        value: 48000
+    LC3_BITRATE:
+        description: LC3 bitrate
+        value: 96000
+
+    BIG_PHY:
+        description:
+        value: 2
+    BIG_NUM_BIS:
+        description: Max number of BISes used
+        value: 2
+    BIG_RTN:
+        description: BIG RTN (number of retransmissions)
+        value: 0
+    BIG_PACKING:
+        description: Arrangement of BIS subevents (0 = sequential, 1 = interleaved)
+        value: 0
+    BIG_FRAMING:
+        description:
+        value: 0
+    BIG_ENCRYPTION:
+        description: BIS encryption
+        value: 0
 
 syscfg.vals:
     CONSOLE_IMPLEMENTATION: full
@@ -65,3 +116,34 @@ syscfg.vals:
     BLE_ISO_MAX_BISES: 2
 
     BLE_AUDIO: 1
+
+    MCU_HFCLK_DIV: 1
+    HARDFLOAT: 1
+
+syscfg.vals.AUDIO_USB:
+    # TinyUSB
+    USBD_VID: 0xFFFF
+    USBD_PID: 0x0001
+    USBD_VENDOR_STRING: '"Apache Software Foundation"'
+    USBD_PRODUCT_STRING: '"NimBLE Auracast"'
+    USBD_AUDIO_OUT: 1
+    USBD_STACK_SIZE: 500
+    USBD_STD_DESCRIPTORS: 0
+
+    USB_AUDIO_OUT_CHANNELS: MYNEWT_VAL(AURACAST_CHAN_NUM)
+
+    # Resampler
+    LIBSAMPLERATE_ENABLE_SINC_BEST_CONVERTER: 0
+    LIBSAMPLERATE_ENABLE_SINC_MEDIUM_CONVERTER: 0
+    LIBSAMPLERATE_ENABLE_SINC_FAST_CONVERTER: 1
+    LIBSAMPLERATE_LIBSAMPLER_NDEBUG: 1
+
+syscfg.vals.ISO_HCI_FEEDBACK:
+    BLE_LL_ISO_HCI_FEEDBACK_INTERVAL_MS: 1000
+    BLE_HS_GAP_UNHANDLED_HCI_EVENT: 1
+
+syscfg.vals.BSP_NRF52:
+    BLE_PHY_NRF52_HEADERMASK_WORKAROUND: 1
+
+$import:
+    - "@apache-mynewt-nimble/apps/auracast/syscfg.usb.yml"
