--- apps/auracast_usb/pkg.yml
@@ -1,46 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#  http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-#
-
-pkg.name: apps/auracast_usb
-pkg.type: app
-pkg.description: Auracast sample application.
-
-pkg.author: "Krzysztof Kopyściński"
-pkg.email: "krzysztof.kopyscinski@codecoup.pl"
-pkg.homepage: "http://mynewt.apache.org/"
-pkg.keywords:
-
-pkg.deps:
-    - "@apache-mynewt-core/sys/config"
-    - nimble/host
-    - nimble/host/util
-    - nimble/host/services/gap
-    - nimble/host/store/config
-    - "@apache-mynewt-core/kernel/os"
-    - "@apache-mynewt-core/sys/console"
-    - "@apache-mynewt-core/sys/log"
-    - "@apache-mynewt-core/sys/stats"
-    - "@apache-mynewt-core/sys/sysinit"
-    - "@apache-mynewt-core/sys/id"
-    - "@apache-mynewt-core/hw/usb/tinyusb"
-    - "@apache-mynewt-nimble/nimble/host/audio/services/auracast"
-    - "@apache-mynewt-nimble/ext/liblc3"
-    - "@apache-mynewt-nimble/ext/libsamplerate"
-
-pkg.init:
-    audio_usb_init: 402
